phased_barrier: move gate allocation to start
Currently, advance_and_wait() allocated a new gate
which might fail.

In order to make it fail-free, move the allocation to
start() that will allocate a new lw_shared gate if the
current one is disengaged.

Use a shared_future for advance_and_wait so that
subsequent calls to advance_and_wait while one
is already pending will join waiting on the previous gate.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>