compaction: propagate error via reader.abort
When compaction is stopped reader.consume_in_thread will throw
compaction_stop_exception.  Propagate it to the reader via
reader.abort() so that it could propagate the error
via the queue_reader_handle.

It's needed here because without it, the bucket writers
get either a dangling reader error or a broken_promise (because
they might wait on the queue_reader _full or _not_full promises
that get destroyed later.  Explicitly aborting the queue reader here
allows it to set the exception on these promises and on the
queue_reader_handle so it is passed correctly up the chain.

Eventually the compaction_stop_exception is caught and handled by
compaction_manager::run_custom_job that just logs it rather than
propagating a failure.

Fixes #8755

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>