flat_mutation_reader: consider timeout when consuming reader
The timeout parameter is passed down to the reader implementations
fill_buffer and friends, but it is ignored there.

Consider it in a generic way so to throw a timeout_error
if the timeout has expired, right before calling fill_buffer.

If the timeout expires during or after fill_buffer
it will not be considered and any mutations already in the
buffer will be available.  Same goes for end_fo_stream,
if that happened before a the timeout is reconsidered,
the reader may be consumed normally with no error.

This change is good as a safety net. Implementations should still
be fixed to respect the timeout. A single fill_buffer() may take
a long time.

Add a unit test reproducing the issue and validating the fix.

Fixes #8846

Test: unit(dev), sstable_reader_timeout_test(debug)
Signed-off-by: Benny Halevy <bhalevy@scylladb.com>