test.py: create log file in tmpdir/mode for a single mode
test.py is automated in Jenkins where it is executed
separately for each mode.  In this case, the different instances
creating the log file in the same location - the root of the default
temporary directory (./testlog) - are overwriting each other.

Simply create the log file in the <tmpdir>/<mode> directory
if a single mode is given to separate the logs.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>