locator: add effective_replication_map_registry
Add a compound key to effective_replication_map for uniquely identify
it in the effective_replication_map_registry.

Start a sharded effective_replication_map_registry in main
and pass a reference to it to storage_proxy and storage_server.

For each keyspace, calculate the registry key using the replication
strategy and current token_metadata_ptr and use it to lookup
the registry.

If found in the registry, just keep a shared pointer to
the effective_replication_map, otherwise construct
a new one and insert it to the registry.

When registered, the effective_replication_map unregisters
itself when destroyed, and then it schedules an async task
to clear_gently its contents, protected by the registry gate.
The gate is closed in effective_replication_map_registry::stop.

Note that for non-shard 0 instances, if the map
is not found in the registry, we construct it
by cloning the precalculated replication_map
from shard 0 to save the cpu cycles of re-calculating
it time and again on every shard.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>