keyspace, storage_service: use effective_replication_map registry to share across keyspaces
Establish an effective_replication_map_registry per database shard.
For each keyspace, calculate the registry key using the replication
strategy and current token_metadata_ptr and use it to lookup
the registry.

If found in the registry just keep a shared pointer to
the effective_replication_map_ptr, otherwise construct
a new one and insert it to the registry.

Note that for non-shard 0 instances, if the map
is not foun in the registry, we construct it
by cloning the precalculated replication_map
from shard 0 to save the cpu cycles of re-calculating
it time and again on every shard.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>