service: add services_controller
base_controller and services_controller on top of it
provide infrastructure for keeping the inter-service
dependencies and then to do bottom-up or top-down
traversals of all services for performing high-level
start, serve, drain, shutdown, and stop functions.

When starting the services, low-level services
are started first, then the service that depend on them.
For drain, shutdown, and stop, the higher-level services
are handled first, then their dependencies.

This simplifies the handling of failed or aborted start
as we keep track of each service state and stop
on in the error handling path if needed.

A unit test was added to test this functionality
but it should eventually be implemented
in main and cql_test_end.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>