compaction_manager: compaction_state: use counter for compaction_disabled
We;d like to use compaction_state::gate both for functions
running with compaction disabled and for and task referring
to the compaction_state so that stop_ongoping_compaction
can wait on all functions referring to the state structure.

This is also cleaner with respect to not relying on
gate::use_count() when re-submitting regular compaction
when compaction is re-enabled.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>