Add service_ctl
service_ctl is a framework for controlling
service lifecycle states: start, serve, drain,
shutdown, and stop, and the dependencies
between services, such that the respective state
transitions will happen in the right order.

Currently starting and stopping all services
is done linearly, roughly relying on the order
of the start() calls to start low-level service
before high level ones, and then we rely on the
reverse order of definition and on explicit
deferred action to stop and destroy the service -
which breaks quite often during shutdown
when lower-level services shut down while
they are needed for orderly shutting down
higher-level services that use them.

main, cql_test_env, and the likes will be converted
to use the service_ctl sub system in the following
patches to define service controllers and their
state actions, and the dependencies between them.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>