utils: add job registry
Jobs are designed to track long running actions in scylla
like repair, compaction, and node operations.

A job can be used to request abort of the action
and to wait for its completion.

A job is identified either by its uuid or
by per-module dequential job-numbers.
Both are auto-generated by the job registry
or they can be given in the job config, when created.

When, created, a retention time can be configured
to keep the job object for some while afte it
compeltes so its status can be retrieved bny the
api layer even after it finished.

The job_registry is used to keep track of all jobs,
list them and find jobs based on their uuid or job_number.

job_registry::stop should be called before destroying
the registry for aborting and waiting for all background jobs.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>