repair_service: add repair_tracker getter
And rename the global repair_tracker getter to
`the_repair_tracker` as the first step to get rid of it.

repair_service methods now use the repair_service::repair_tracker
method.

The global getter was renamed to `the_repair_tracker()`
temporarily while eliminating it in this series
to help distinguish it from repair_service::repair_tracker().

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>