lister: add queue_lister
queue_lister provides a simpler interface
compared to lister.

After creating the queue_lister,
its async get() method should be called repeatedly,
returning a std::optional<directory_entry> each call,
until it returns a disengaged entry or an error.

This is especially suitable for coroutines
as demonstrated in the unit tests that were added.

For example:

    auto ql = queue_lister(path);
    while (auto de = co_await ql.get()) {
        co_await process(*de);
    }

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>