table: add perform_offstrategy_compaction
Expose an async method to perform offstrategy- compaction, if needed.

Returns a future<bool> that is resolved when offstrategy_compaction completes.
The future value is true iff offstrategy compaction was required.

To be used in a following patch by the storage_service api.

Call it from `trigger_offstrategy_compaction` that triggers
offstrategy compaction in the background and warn about ignored
failures.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>