replica: table: add static snapshot_on_all method
Since the snapshot operation on a table is sharded,
make table::snapshot a static method (renamed
to snapshot_on_all) that accepts a global_table_ptr
and coordinates the snapshot operation on all shards.

The main benefit of doing so is making sure that
the snapshot_manager object is created first, before
the actual snapshotting begins on any shard,
and making sure it is cleaned up properly, even if any
of the shards failed.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>