types: time_point_to_string: harden against out of range timestamps
The time point is multiplied by an adjustment factor of 1000
for boost::posix_time::time_duration::ticks_per_second() = 1000000
when calling boost::posix_time::milliseconds(count)
and that may lead to integer overflow as reported
by the UndefinedBehaviorSanitizer.

See https://github.com/scylladb/scylla/issues/10830#issuecomment-1158899187

This change used the fmt library to format the iso representation
of the time stamp to avoid exceptions and error handling.

Note that time points with large durations will end up
very far in the future (or in the past if they are negative).

Refs #10830

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>