range_tombstone_change_generator: update _lower_bound on early return when empty
With change 18a80a98b870c1104295fc48a31e2e4b5089ef3a flush
returns early if _range_tombstones.empty.

This missed a side effect of the function of
setting _lower_bound to upper_bound, that happens
without the optimization.

Although it seems to be benign in 5.1.dev,
it proved to be required for backporting to 5.0
(See https://github.com/scylladb/scylla/pull/10969)
so let's add it to 5.1.dev as well.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>