abstract_replication_strategy: add maybe_calculate_natural_endpoints
Call the strategy's calculate_natural_endpoints()
only once if it has_static_natural_endpoints()
(like local_strategy or everywhere_replication_strategy)

Use the token_metadata ring version to detect if the
token_metadata might have changed (and to detect the first time
around) in order to recalculate the natural endpoints -
although this shouldn't be the case with the present users.
We still want to do that since `maybe_calculate_natural_endpoints`
is defined as a public method and external users
might call it with different versions of token_metadata.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>