logalloc: tracker::impl: occupancy: get rid of reclaiming_lock
It was added in d20fae96a2da5d0fe381c91c42be6253503eb795
as a precaution not to invalidate iterators while
traversing _regions.  However it is not requried as no allocation
is done on this synchronous path - therefore there is no
point in preventing reclaim.

This will allow making the respective functions const
as they merely return stats and do not modify the tracker impl.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>