token_metadata: count_normal_token_owners: use topology locations
We don't need to calculate the unique set of normal token
owners, nor do we have to maintain a member holding it.

It is already available in the topolgy locations map
as the set of (unique) keys.

This set is already incrementally updated in
update_normal_tokens and in remove_endpoint
and it is already gently cloned by clone_only_token_map
and cleared by clear_gently.

Fixes #11128

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>