utils: add UUID_class and use for counter_id
We would like to define more distinct types
that are currently using plain UUID to identify
resources in the system, like table id and schema
version id.

As with counter_id, the motivation is to restrict
he usage of the distinct types so they can be used
(assigned, compared, etc.) only with objects of
the same type.  Using with a generic UUID will
then require explicit conversion, that we want to
expose.

It would be nice if we could just inherit utils::UUID,
or make utils::UUID derive from UUID_class<void>
but this doesn't work in a simple way since
UUID is not trivially default constructible, and we
require that for harnessing the serialization boiler plate.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>