mutation_compactor: compact_mutation_state: get optional reference to compaction_manager
So it can be used for get_gc_before_for_key
in a following patch.

In general, we pass std::nullopt from the query path
and the reference to the compaction_manager when compacting sstables
so we can use it for tombstone gc purging, based on the repair_history
timestamp in repair gc mode.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>