frozen_mutation: frozen_mutation_consumer_adaptor: always consume flushed mutation_fragments
Always consume the flushed mutation_fragments
and set _stop_consuming if needed, to be returned later.
We must not drop the mutation fragments even
when stop_iteration::yes is returned fomr the consumer.

Note that unlike the mutation counterpart,
frozen_mutation::consume functions aren't pausable
and don't return a cookie, so the utility of returning
stop_iteration::yes from the consumer is limited.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>