cql3: attributes: introduce mask and get_set()
Define a bitmask for each type of attribute:
timestamp, time_to_live, and timeout, plus
respective operators (&, |, ~) for turning
them into an `unsigned` bitmask.

And add a get_set() method to attributes
and attributes::raw that returns the set
attributes as a bitmask.

Use that for verifying the attributes in
select and delete statements that allow
USING only a subset of the attributes.

A following patch, will also implement those
subsets at the grammar level rather than
verifying them in the respective statement's
constructors.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>