db/large_data_handler: cql_table_large_data_handler: record_large_cells: record large_collections
When the large_collection_detection cluster feature is enabled,
select the internal_record_large_cells_and_collections method
to record the large collection cell, storing also the collection_items
column.

We want to do that only when the cluster feature is enabled
to facilitate rollback in case rolling upgrade is aborted,
otherwise system.large_cells won't be backward compatible
and will have to be deleted manually.

Closes #11449

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>