system_keyspace: add quarantined_hosts table
To be used for keeping a record of hosts
that were remved from the cluster and must not join back.

Each quarantined host is stored as a row in a single
partition. A `quarantined_at` timestamp is stored along with
(as milliseconds since the unix epoch)

Inserts are done with TTL of 30 days
so not to accumulate quarantined hosts forever.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>